Use a timestamp with nanoseconds
For multi-file benchmarks, many log agents use some sort of logic to
determine if a file is a duplicate or is rotated. With second-resolution
timestamps, the beginnings of files are usually identical, so the files
are often marked as duplicates by the log agent and skipped, making the
benchmarks inaccurate.

This commit changes the timestamp to one which includes nanoseconds to
mitigate that issue.